pyproject.toml
[project]
name = "hello"
version = "0.1.0"
description = "An example Django app running in Docker."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django==6.0.3",
  "celery==5.6.3",
  "django-debug-toolbar==6.3.0",
  "djangorestframework==3.16.0",
  "django-filter==25.1",
  "gunicorn==25.3.0",
  "psycopg==3.3.3",
  "redis==7.4.0",
  "ruff==0.15.9",
  "setuptools==82.0.1",
  "whitenoise==6.12.0",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["I", "SIM"]
